Added a hint of "datetime" to allow easier discovery and mapping.
Currently searching for anything related to a "date" in the IPSO Smart Objects gives 0 results.  I have added additional description to help assist with searchability in finding this object within IPSO for the concept of a "datetime" that we expressed as Unix Time.  This will help later with mapping to other ontologies or vocabularies.